Rename: user_validator 파일 user_utils에 병합, user_utils 파일 utils 폴더로 이동
