+++ mcp-api-backend/api/v1/adminInfo.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 from db.connection import get_db
-from service.user import user_service
+from service import user_service
 
 router = APIRouter()
