+++ mcp-api-backend/service/user_service.py
@@ -10,7 +10,7 @@ from sqlalchemy.orm import Session
 from db.connection import get_db
 from utils.utils import object_as_dict
 from entity.user_entity import UserInit, UserCreate, User
-from service.user import user_util
+from utils.user_utils import validate_password
 from src.shared.security import deps
 
 def get_admin_info(db: Session):
@@ -20,7 +20,7 @@ def get_admin_info(db: Session):
 
 async def create_init_user(passwd: UserInit, db: Session = Depends(get_db)):
     init_user = settings.INIT_USER
-    user_util.validate_password(init_user.get("username"), passwd.password)
+    validate_password(init_user.get("username"), passwd.password)
     db_user = crud_users.get_user_by_username(db, username=init_user.get("username"))
     if db_user:
         raise HTTPException(status_code=409, detail="Username already registered")
@@ -39,7 +39,7 @@ async def create_user(
     db_user = crud_users.get_user_by_username(db, username=user.username)
     if db_user:
         raise HTTPException(status_code=400, detail="Username already registered")
-    user_util.validate_password(user.username, user.password)
+    validate_password(user.username, user.password)
     try:
         result = crud_users.create_user(db=db, user=user)
         # TODO: logging 추가
@@ -85,7 +85,7 @@ async def update_user(
     # TODO: user 권한 validation
     check_None = [None, "", "string"]
     if user.password not in check_None:
-        user_util.validate_password(user.username, user.password)
+        validate_password(user.username, user.password)
     try:
         result = crud_users.update_user(db=db, user_id=user_id, user=user)
         # TODO: logging 추가
