+++ mcp-api-backend/utils/user_utils.py
@@ -1,3 +1,6 @@
+from dependency_injector.wiring import Provide, inject
+from fastapi import Depends, HTTPException, status
+
 from typing import Tuple
 
 from dependency_injector import containers, providers
@@ -56,3 +59,20 @@ class Container(containers.DeclarativeContainer):
     user_validate_service = providers.Singleton(
         UserValidator, username_validator, password_validator
     )
+
+@inject
+def validate_password(
+    username: str, password: str, user_validate_service=Provide[Container.user_validate_service]
+):
+    (result, additional_info) = user_validate_service.validate(username, password)
+    print((result, additional_info))
+    if not result:
+        raise HTTPException(
+            status_code=400,
+            detail=f"{additional_info}",
+        )
+    return True
+
+
+container = Container()
+container.wire(modules=[__name__])
